ifupdown: switch static and stateless configuration to ifupdown
While not nearly as exciting, ifupdown provides a significantly cleaner
interface for static network configuratons. On Void it has the
additional benefit of being run as part of stage 1 which guarantees a
static network before the supervisor starts bringing up other services
which should have additional benefits around service reliability. I
tried to make things as straight forward an unsurprising as possible so
many of the ansible_default_* calls have been removed. This means that
host layouts like b-hel-fi will most likely be incorrect until a
network_interfaces block can be written but there should be less
reliance on prior state to correctly configure a system.

If races still exist on the DO hosts between dhcp resolution and network
clients there are a couple of ways to solve that. The simplest is to add
some variety of DNS call to short-circuit the dependent run scripts
though if that isn't good enough I have an entirely experimental branch
that emulates service readiness with dhcpcd and runit.

Caveats:
This commit does not entirely shut off dhcpcd for fully static builds
however. The current builds rely on dhcpcd to handle merging the
configuration and it was easier to borrow that behavior for the time
being. Also, I didn't have access to running configs and my VM test
system needs some love so I wasn't able to end-to-end satisfactorily
test every configuration (notably the pure DHCP and SLAAC setups). I'm
pretty sure it covers all of the main cases but I wouldn't be surprised
if bugs are hiding out somewhere.